Unify delete messages api

+++ server/console_channel.go
@@ -11,9 +11,10 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/emptypb"
 	"net/url"
 	"sort"
+	"strconv"
+	"strings"
 	"time"
 )
 
@@ -45,39 +46,51 @@ func (s *ConsoleServer) ListChannelMessages(ctx context.Context, in *console.Lis
 	return messageList, nil
 }
 
-func (s *ConsoleServer) DeleteChannelMessage(ctx context.Context, in *console.DeleteChannelMessageRequest) (*emptypb.Empty, error) {
-	messageID, err := uuid.FromString(in.Id)
-	if err != nil {
-		return nil, status.Error(codes.InvalidArgument, "Requires a valid message ID.")
-	}
-
-	query := "DELETE FROM message WHERE id = $1::UUID"
-	if _, err := s.db.ExecContext(ctx, query, messageID); err != nil {
-		s.logger.Error("Could not delete message.", zap.Error(err))
-		return nil, status.Error(codes.Internal, "An error occurred while trying to delete the message.")
-	}
-
-	s.logger.Info("Message deleted.", zap.String("message_id", messageID.String()))
-	return &emptypb.Empty{}, nil
-}
-
 func (s *ConsoleServer) DeleteChannelMessages(ctx context.Context, in *console.DeleteChannelMessagesRequest) (*console.DeleteChannelMessagesResponse, error) {
-	query := "DELETE FROM message WHERE create_time < $1::TIMESTAMPTZ"
-	deleteBefore := time.Unix(in.Before.Seconds, int64(in.Before.Nanos)).UTC()
+	var affected int64
+	if in.Before != nil {
+		query := "DELETE FROM message WHERE create_time < $1::TIMESTAMPTZ"
+		deleteBefore := time.Unix(in.Before.Seconds, int64(in.Before.Nanos)).UTC()
 
-	var res sql.Result
-	var err error
-	if res, err = s.db.ExecContext(ctx, query, &pgtype.Timestamptz{Time: deleteBefore, Status: pgtype.Present}); err != nil {
-		s.logger.Error("Could not delete messages.", zap.Error(err))
-		return nil, status.Error(codes.Internal, "An error occurred while trying to delete old message.")
+		var res sql.Result
+		var err error
+		if res, err = s.db.ExecContext(ctx, query, &pgtype.Timestamptz{Time: deleteBefore, Status: pgtype.Present}); err != nil {
+			s.logger.Error("Could not delete messages.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "An error occurred while trying to delete messages.")
+		}
+		affected, err = res.RowsAffected()
+		if err != nil {
+			s.logger.Error("Could not count deleted messages.", zap.Error(err))
+		}
+		s.logger.Info("Messages deleted.", zap.Int64("affected", affected), zap.String("timestamp", deleteBefore.String()))
 	}
-	affected, err := res.RowsAffected()
-	if err != nil {
-		s.logger.Error("Could not count deleted messages.", zap.Error(err))
-		return &console.DeleteChannelMessagesResponse{Total: 0}, nil
+	if len(in.Ids) > 0 {
+		params := make([]interface{}, 0, len(in.Ids))
+		statements := make([]string, len(in.Ids))
+		for i, id := range in.Ids {
+			idStr, err := uuid.FromString(id)
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, "Requires a valid message ID.")
+			}
+			params = append(params, idStr)
+			statements[i] = "$" + strconv.Itoa(i)
+		}
+		query := "DELETE FROM message WHERE id IN (" + strings.Join(statements, ",") + ")"
+		var res sql.Result
+		var err error
+		if res, err = s.db.ExecContext(ctx, query, params...); err != nil {
+			s.logger.Error("Could not delete message.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "An error occurred while trying to delete batch messages.")
+		}
+		batch, err := res.RowsAffected()
+		if err != nil {
+			s.logger.Error("Could not count deleted messages.", zap.Error(err))
+		}
+		affected = affected + batch
+
+		s.logger.Info("Messages batch deleted.", zap.Int("size", len(in.Ids)))
 	}
 
-	s.logger.Info("Messages deleted.", zap.Int64("affected", affected), zap.String("timestamp", deleteBefore.String()))
 	return &console.DeleteChannelMessagesResponse{Total: affected}, nil
 }
